ISPN-7184 Server startup can fail after the upgrade to JGroups 4
Re-add support for empty messages as a signal for missing fork channels
both in embedded mode and in server mode.

The server uses an empty message to signal that the fork channel with
that name is not running on a remote node.
ISPN-6970 made some changes to response unmarshalling and broke that.